Image Watermarking App/requirements.txt
flask
opencv-python-headless==4.2.0.32
Pillow
scikit-image
numpy
random2
strings